training_service/requirements.txt
fastapi
uvicorn[standard]
asyncpg
sqlalchemy
psycopg2-binary
keras
opencv-python
dlib
pytz